requirements.txt
altair==4.2.0
appdirs==1.4.4
attrs==22.1.0
audioread==3.0.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
gitdb==4.0.9
GitPython==3.1.29
htbuilder==0.6.0
idna==3.4
importlib-metadata==5.0.0
iteration-utilities==0.11.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
llvmlite==0.39.1
MarkupSafe==2.1.1
numba==0.56.3
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
pooch==1.6.0
protobuf==3.20.3
pyarrow==9.0.0
pycparser==2.21
pydeck==0.8.0b4
pydub==0.25.1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytube @ git+https://github.com/pytube/pytube@84faec34c8a66f502ac635a5610445dbff160654
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
resampy==0.4.2
rich==12.6.0
scikit-learn==1.1.3
scipy==1.9.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
soundfile==0.11.0
st-annotated-text==3.0.0
streamlit==1.13.0
streamlit-ext==0.1.4
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
youtube-transcript-api==0.5.0
zipp==3.10.0
